Move blacklist logic to separate module
- shouldn't have been in activity-logger as it's grown out of that
