+++ src/activity-logger/index.js
@@ -4,7 +4,6 @@ import docuri from 'docuri'
 import randomString from 'src/util/random-string'
 
 export const visitKeyPrefix = 'visit/'
-export const blacklistStorageKey = 'blacklist'
 
 export const visitDocsSelector = { _id: { $gte: visitKeyPrefix, $lte: `${visitKeyPrefix}\uffff` } }
 
@@ -23,70 +22,4 @@ export function generateVisitDocId({timestamp, nonce} = {}) {
         timestamp: date.getTime(),
         nonce: nonce || randomString(),
     })
-}
-
-/**
- * Given a URL and user's blacklist, checks the URL against the blacklist expressions to see if any
- * rule matches it.
- *
- * @param {string} url The URL to check against the blacklist.
- * @param {Array<any>} blacklist Blacklist data to check URL against.
- * @return {boolean} Denotes whether or not the given URL matches any blacklist expressions.
- */
-function isURLBlacklisted(url = '', blacklist = []) {
-    // Main checking logic between a given blacklist expression and current URL
-    //   TODO: make this more "smart" (maybe check parts of the URL instead of match all, for e.g.)
-    const doesExpressionMatchURL = (expression = '') => url.includes(expression)
-
-    // Reduces blacklist to a bool by running main checking logic against each blacklist expression
-    // (returns true if a single match is found in entire blacklist)
-    return blacklist.reduce((prev, curr) => doesExpressionMatchURL(curr.expression) || prev, false)
-}
-
-// Tell whether a page can be stored.
-export function isLoggable(url) {
-    // Only http(s) pages. Ignoring data URLs, newtab, ...
-    const loggableUrlPattern = /^https?:\/\//
-    return loggableUrlPattern.test(url)
-}
-
-/**
- * Fetches ready-to-use blacklist array from storage.
- * @returns {Array<any>} Blacklist represented by nn array of objects containing `expression` keys
- */
-async function fetchBlacklist() {
-    // Fetch and parse blacklist data for check calls to use
-    const { blacklist } = await browser.storage.local.get(blacklistStorageKey)
-
-    return !blacklist ? [] : JSON.parse(blacklist)
-}
-
-async function storeBlacklist(blacklist = []) {
-    const serialized = JSON.stringify(blacklist)
-
-    await browser.storage.local.set({ [blacklistStorageKey]: serialized })
-}
-
-/**
- * HOF which wraps a blacklist checking function up with stored blacklist data.
- * @returns {({ url: string }) => boolean} Ready-to-use checking function against a URL
- */
-export async function checkWithBlacklist() {
-    const blacklist = await fetchBlacklist()
-
-    /**
-     * @param {string} url The URL to check against the blacklist
-     */
-    return ({ url = '' } = {}) => isLoggable(url) && !isURLBlacklisted(url, blacklist)
-}
-
-const createBlacklistEntry = url => ({
-    expression: url,
-    dateAdded: Date.now(),
-})
-
-export async function addToBlacklist(url) {
-    const blacklist = await fetchBlacklist()
-
-    await storeBlacklist([...blacklist, createBlacklistEntry(url)])
 }
