+++ src/popup/container.js
@@ -2,12 +2,12 @@ import React, { Component } from 'react'
 import qs from 'query-string'
 
 import extractTimeFiltersFromQuery from 'src/util/nlp-time-filter.js'
-import { checkWithBlacklist, addToBlacklist } from 'src/activity-logger'
+import { checkWithBlacklist, addToBlacklist } from 'src/blacklist'
+import { getPageDocId, updateArchiveFlag } from './archive-button'
 import Popup from './components/Popup'
 import Button from './components/Button'
 import LinkButton from './components/LinkButton'
 import SplitButton from './components/SplitButton'
-import { getPageDocId, updateArchiveFlag } from './archive-button'
 
 export const overviewURL = '/overview/overview.html'
 export const optionsURL = '/options/options.html'
